requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-phonenumber-field==6.0.0
django-postgres-extensions==0.9.3
django-utils-six==2.0
djangorestframework==3.12.4
drf-extra-fields==3.2.1
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
importlib-metadata==4.8.2
Markdown==3.3.6
phonenumbers==8.12.38
psycopg2==2.9.2
psycopg2-binary==2.8.5
pytz==2021.3
six==1.15.0
sqlparse==0.4.2
whitenoise==5.3.0
zipp==3.6.0
